rocky update install.sh

+++ install.sh
@@ -1,5 +1,7 @@
-ln -s $HOME/dot_config $HOME/.config
-ln -s $HOME/dot_config/.zshenv $HOME/.zshenv
+if [[ ! -e $HOME/.config ]]; then
+	ln -s $HOME/dot_config $HOME/.config
+	ln -s $HOME/dot_config/.zshenv $HOME/.zshenv
+fi
 
 pip3 install requests tqdm sh pynvim
 python3 $HOME/.config/scripts/install.py
